movie-recommender/pyproject.toml
[project]
name = "movie-recommender"
version = "0.1.0"
description = "Production-style two-stage personalized movie recommender on MovieLens 25M"
requires-python = ">=3.12"
authors = [{ name = "Prasanth Gururaj" }]
readme = "README.md"
license = { text = "MIT" }

dependencies = [
  "pandas>=2.2.0,<3.0.0",
  "numpy>=1.26.0,<2.0.0",
  "pyarrow>=15.0.0,<20.0.0",
  "scipy>=1.13.0,<2.0.0",
  "scikit-learn>=1.5.0,<2.0.0",
  "xgboost>=2.1.0,<3.0.0",
  "lightgbm>=4.4.0,<5.0.0",
  "implicit>=0.7.2,<1.0.0",
  "faiss-cpu>=1.8.0,<2.0.0",
  "mlflow>=2.14.0,<3.0.0",
  "fastapi>=0.111.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "asyncpg>=0.29.0,<1.0.0",
  "redis>=5.0.0,<6.0.0",
  "prometheus-client>=0.20.0,<1.0.0",
  "click>=8.1.0,<9.0.0",
  "pyyaml>=6.0.0,<7.0.0",
  "python-dotenv>=1.0.0,<2.0.0",
  "tqdm>=4.66.0,<5.0.0",
  "python-multipart>=0.0.9,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4.0,<1.0.0",
  "mypy>=1.10.0,<2.0.0",
  "types-pyyaml>=6.0.0",
  "types-redis>=4.6.0",
  "pandas-stubs>=2.2.0",
  "pytest>=8.2.0,<9.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
  "pytest-asyncio>=0.23.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "pip-tools>=7.4.0,<8.0.0",
]

[project.scripts]
train    = "train:train"
evaluate = "evaluate:evaluate"

[build-system]
requires      = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where   = ["."]
include = ["src*"]

[tool.ruff]
line-length    = 88
target-version = "py312"
src            = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "N", "SIM"]
ignore = ["E501", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/*"     = ["S101"]
"train.py"    = ["T201"]
"evaluate.py" = ["T201"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version         = "3.12"
warn_return_any        = true
warn_unused_configs    = true
disallow_untyped_defs  = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["xgboost.*", "lightgbm.*", "faiss.*", "implicit.*", "mlflow.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths    = ["tests"]
asyncio_mode = "auto"
addopts      = "-v --tb=short"
markers = [
  "slow: marks tests as slow",
  "integration: marks integration tests",
]

[tool.coverage.run]
source = ["src"]
omit   = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
fail_under   = 80
show_missing = true